### 2021-11-18 #3922 @whedon
[PRE REVIEW]: PyNumDiff: A Python package for numerical differentiation of noisy time-series data
**Submitting author:** @florisvb (<a href="http://orcid.org/0000-0001-6538-7179">Floris van Breugel</a>)
**Repository:** <a href="https://github.com/florisvb/PyNumDiff" target ="_blank">https://github.com/florisvb/PyNumDiff</a>
**Version:** v0.0.3
