### 2021-11-18 #3922 @arfon
@florisvb – thanks for your submission to JOSS. It looks like the paper can't be compiled because the indentation in the YAML header for this row is off:

```
- name: Department of Biology, University of Washington
   index: 3
```

Could you take a look at fixing this please?
