### 2021-11-18 #3922 @florisvb
Hi Arfon,

I just fixed that issue. Let me know if anything else comes up.

fyi - the paper compiled for me when I ran it through the whedon.theoj.org
link both before and after I made that change, which seems odd.

- Floris

On Thu, Nov 18, 2021 at 4:52 AM Arfon Smith ***@***.***>
wrote:

> @florisvb <https://github.com/florisvb> – thanks for your submission to
> JOSS. It looks like the paper can't be compiled because the indentation in
> the YAML header for this row is off:
>
> - name: Department of Biology, University of Washington
>
>    index: 3
>
>
> Could you take a look at fixing this please?
>
