### 2021-11-19 #3922 @christinahedges
Hello @florisvb, welcome to JOSS! Thanks for your submission. Taking a look at your paper submission I can see a few fixable issues that you may want to work on while I find reviewers:

1. The references are missing DOIs (see whedon's comments above)
2. In your submission we usually encourage a state of the field section or paragraph which references common software that is use to solve similar problems in the field. You should add some references to the tools you mention in your submission.

I will get to work finding some reviewers.
