+++ paper.bib
@@ -5,7 +5,8 @@
   volume={8},
   pages={196865--196877},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/access.2020.3034077},
 }
 
 @incollection{chambolle2010introduction,
@@ -14,7 +15,8 @@
   booktitle={Theoretical foundations and numerical methods for sparse recovery},
   pages={263--340},
   year={2010},
-  publisher={de Gruyter}
+  publisher={de Gruyter},
+  doi={10.1515/9783110226157.263},
 }
 
 @article{brunton2016discovering,
@@ -25,7 +27,8 @@
   number={15},
   pages={3932--3937},
   year={2016},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={10.1073/pnas.1517384113}
 }
 
 @article{de2020pysindy,
@@ -43,7 +46,8 @@
   number={10},
   pages={764--774},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cpc.2007.03.009}
 }
 
 @article{butterworth1930theory,
@@ -75,7 +79,8 @@
   number={4},
   pages={111--117},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/msp.2011.941097}
 }
 
 @article{savitzky1964smoothing,
@@ -86,24 +91,28 @@
   number={8},
   pages={1627--1639},
   year={1964},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/ac60214a047}
 }
 
 @article{kalman1960new,
   title={A new approach to linear filtering and prediction problems},
   author={Kalman, Rudolph Emil},
-  year={1960}
+  year={1960},
+  doi={10.1109/9780470544334.ch9}
 }
 
-@article{henderson2010fundamentals,
-  title={Fundamentals of Kalman Filtering: a Practical Approach--Third edition. Progress in Aeronautics and Astronautics series Vol. 232 P. Zarchan and H. Musoff American Institute of Aeronautics and Astronautics, 1801 Alexander Bell Drive, Suite 500, Reston, VA 20191-4344, USA. 2009. Distributed by Transatlantic Publishers Group, Unit 242, 235 Earls Court Road, London, SW5 9FE, UK.(Tel: 020-7373 2515; e-mail:-richard@ tpgltd. co. uk). 853pp. Illustrated.{\pounds} 100.[10\% discount available to RAeS members on request]. ISBN 978-1-60086-718-7.},
-  author={Henderson, Geoff},
-  journal={The Aeronautical Journal},
-  volume={114},
-  number={1152},
-  pages={121--122},
-  year={2010},
-  publisher={Cambridge University Press}
+@article{henderson2010fundamentals, 
+title={Fundamentals of Kalman Filtering: A Practical Approach}, 
+volume={105}, 
+number={1049}, 
+journal={The Aeronautical Journal}, 
+publisher={Cambridge University Press}, 
+author={Henderson, Geoff.T.}, 
+year={2001}, 
+pages={400–400},
+doi={10.1017/S000192400001232X}, 
+}
 }
 
 @article{aravkin2017generalized,
@@ -113,7 +122,8 @@
   volume={86},
   pages={63--86},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.automatica.2017.08.011}
 }
 
 @book{crassidis2004optimal,
@@ -132,7 +142,8 @@
   number={1-4},
   pages={259--268},
   year={1992},
-  publisher={Elsevier}
[…]
