+++ paper.md
@@ -29,11 +29,15 @@ bibliography: paper.bib
 
 # Statement of need
 
-The numerical computation of derivatives is ubiquitous in every scientific discipline and engineering application because derivatives express fundamental relationships among many quantities of interest. As a result, a large number of diverse algorithms have been developed to differentiate numerical data.  These efforts are challenging because, in reality, practitioners  often have sparse and noisy measurements and data, which undermine the ability to estimate accurate derivatives.  Among the diversity of mathematical approaches that have been formulated, many are ad hoc in nature and require significant bespoke tuning of multiple parameters to produce reasonable results. Thus, at a practical level, it is often unclear which method should be used, how to choose parameters, and how to compare results from different methods.
+The numerical computation of derivatives is ubiquitous in every scientific discipline and engineering application because derivatives express fundamental relationships among many quantities of interest. As a result, a large number of diverse algorithms have been developed to differentiate numerical data.  These efforts are challenging because, in reality, practitioners  often have sparse and noisy measurements and data, which undermine the ability to estimate accurate derivatives.  Among the diversity of mathematical approaches that have been formulated, many are ad hoc in nature and require significant bespoke tuning of multiple parameters to produce reasonable results. Thus, at a practical level, it is often unclear which method should be used, how to choose parameters, and how to compare results from different methods. 
 
-The ability to assess the most appropriate derivative estimation algorithm to use based upon features of the data is of paramount importance.  Currently, practitioners must often implement a number of methods, relying on disparate software packages, before selecting one that is appropriate for their application.  Whereas high-quality data can leverage simple and computationally efficient algorithms (e.g. finite-differences), data that are sparse and noisy requires more sophisticated algorithms that are also more computationally expensive (e.g. total variation regularization).  Regardless of application domain, scientists of various levels of mathematical expertise would benefit from a unified toolbox for differentiation techniques and parameter tuning. To address these needs, we built the open-source package `PyNumDiff`, with two primary goals in mind: (1) to develop a unified source for a diversity of differentiation methods using a common API, and (2) to provide an objective approach for choosing optimal parameters with a single universal hyperparameter (`gamma`) that functions similarly for all differentiation methods[@van2020numerical]. By filling these needs, `PyNumdiff` facilitates easy computations of derivatives on diverse time-series data sets.
+Regardless of application domain, scientists of various levels of mathematical expertise would benefit from a unified toolbox for differentiation techniques and parameter tuning. To address these needs, we built the open-source package `PyNumDiff`, with two primary goals in mind: (1) to develop a unified source for a diversity of differentiation methods using a common API, and (2) to provide an objective approach for choosing optimal parameters with a single universal hyperparameter (`gamma`) that functions similarly for all differentiation methods[@van2020numerical]. By filling these needs, `PyNumdiff` facilitates easy computations of derivatives on diverse time-series data sets.
+
+# State of the field
+
[…]
