### 2021-11-23 #3922 @florisvb
Sorry for the delay --

I have added in DOIs based on whedon's comments, as well as a state of the
field section with some additional software references.

- Floris

On Fri, Nov 19, 2021 at 9:28 AM Christina Hedges ***@***.***>
wrote:

> Hello @florisvb <https://github.com/florisvb>, welcome to JOSS! Thanks
> for your submission. Taking a look at your paper submission I can see a few
> fixable issues that you may want to work on while I find reviewers:
>
>    1. The references are missing DOIs (see whedon's comments above)
>    2. In your submission we usually encourage a state of the field
>    section or paragraph which references common software that is use to solve
>    similar problems in the field. You should add some references to the tools
>    you mention in your submission.
>
> I will get to work finding some reviewers.
>
