### 2021-11-24 #3922 @christinahedges
@florisvb, thank you for fixing your DOIs (it looks like there's one invalid DOI that is easily fixed.)

I see from your submission that you have [published a paper](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7899139/) on this topic in another journal. There may be some concern from JOSS that this is not a unique submission. Could you please justify that this submission to JOSS is unique, and separate from the previously published article? I will then discuss with the rest of the JOSS team.
