### 2021-11-24 c658647
Update paper.bib
removed doi.org prefixes
