+++ paper.bib
@@ -191,7 +191,7 @@ doi={10.1017/S000192400001232X},
  pages         = {357--362},
  doi           = {10.1038/s41586-020-2649-2},
  publisher     = {Springer Science and Business Media {LLC}},
- url           = {https://doi.org/10.1038/s41586-020-2649-2}
+ url           = {10.1038/s41586-020-2649-2}
 }
 
 @article{demo18pydmd,
@@ -202,7 +202,7 @@ doi={10.1017/S000192400001232X},
   Number = {22},
   Pages = {530},
   Year = {2018},
-  Doi = {https://doi.org/10.21105/joss.00530}
+  Doi = {10.21105/joss.00530}
 }
 
 @Misc{findiff,
