### 2021-11-24 #3922 @florisvb
I fixed the DOI issue.

Regarding the previous IEEE paper. The previous paper was a scientific
advance in optimization methods for numerical differentiation, which we
illustrated using 4 differentiation methods. Our PyNumDiff package, on the
other hand, provides 22 differentiation methods, organized documentation,
and is compliant with all the open source software requirements. Together,
this represents a substantial effort beyond the original scientific
advance. These improvements were made possible by substantial new
contributions from an additional author, Yuying Liu, who was not a part of
the original author list for the scientific advance.

A comparable example would be this scientific paper,
https://www.pnas.org/content/113/15/3932.short, which was followed up by
this JOSS paper, https://joss.theoj.org/papers/10.21105/joss.02104.

Please let me know if I can expand on any of these details.

Cheers,

- Floris

On Wed, Nov 24, 2021 at 9:55 AM Christina Hedges ***@***.***>
wrote:

> @florisvb <https://github.com/florisvb>, thank you for fixing your DOIs
> (it looks like there's one invalid DOI that is easily fixed.)
>
> I see from your submission that you have published a paper
> <https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7899139/> on this topic in
> another journal. There may be some concern from JOSS that this is not a
> unique submission. Could you please justify that this submission to JOSS is
> unique, and separate from the previously published article? I will then
> discuss with the rest of the JOSS team.
>
