### 2021-11-30 #3922 @florisvb
Great! Here are some potential reviewers.

Frank Noe, FU Berlin, ***@***.***
Benjamin Peherstorfer, Courant Institute NYU, ***@***.***
Peter Benner, Magdeburg  ***@***.***
Hayden Schaeffer, Carnegie Mellon, ***@***.***

- Floris

On Tue, Nov 30, 2021 at 9:53 AM Christina Hedges ***@***.***>
wrote:

> Hi @florisvb <https://github.com/florisvb>, I spoke with the editorial
> team, and we agree this is acceptable. I am looking for reviewers for this
> work, please would you let me know your suggestions for anyone in your
> field who is not conflicted with this work, and able to give an impartial
> review.
>
