### 2021-12-22 79bc76e
updated PI control example to be more illustrative and linear
