+++ pynumdiff/utils/_pi_cruise_control.py
@@ -1,195 +1,11 @@
-"""
-Simulation and Control of a cruise ?
-"""
-
-import numpy as np
-from pynumdiff.utils import utility
-
-parameters = {'omega_m': 420,
-              'T_m': 190,
-              'beta': 0.4,
-              'Cr': 0.01,
-              'Cd': 0.32,
-              'A': 2.4,
-              'g': 9.8,
-              'm': 3000,
-              'rho': 1.3,
-              'v_r': 30,
-              'k_p': 2,
-              'k_i': 2
-              }
-
-
-def triangle(iterations, dt):
-    """
-    Create sawtooth pattern of hills for the car to drive over.
-
-    :param iterations: number of time points in time series
-    :type iterations: int
-
-    :param dt: time step in seconds
-    :type dt: float
-
-    :return: time series of hills (angle of road as a function of time)
-    :rtype: np.matrix
-
-    """
-    t = np.arange(0, iterations*dt, dt)
-    continuous_x = np.sin(0.02*t*np.sqrt(t))
-
-    # find peaks and valleys
-    peaks, valleys = utility.peakdet(continuous_x, 0.1)
-
-    # organize peaks and valleys
-    if len(peaks) > 0:
-        reversal_idxs = peaks[:, 0].astype(int).tolist()
-        reversal_vals = peaks[:, 1].tolist()
-    else:
-        reversal_idxs = []
-        reversal_vals = []
-    if len(valleys) > 0:
-        reversal_idxs.extend(valleys[:, 0].astype(int).tolist())
-        reversal_vals.extend(valleys[:, 1].tolist())
-
-    reversal_idxs.extend([0, len(continuous_x)-1])
-    reversal_vals.extend([0, continuous_x[-1]])
-
-    idx = np.argsort(reversal_idxs)
-    reversal_idxs = np.array(reversal_idxs)[idx]
-    reversal_vals = np.array(reversal_vals)[idx]
-    reversal_ts = t[reversal_idxs]
-
-    x = np.interp(t, reversal_ts, reversal_vals)
-    x = np.matrix(x)
-
-    return x
-
-
-def effective_wheel_radius(v=20):
-    """
-    Allow effective wheel radius to be a function of velocity.
-
-    :param v: ignored for now
-    :type v:
-
-    :return: effective wheel radius, constant for now
-    :rtype: float
-    """
-    return v
-
-
-def torque(omega):
-    """
-    Convert throttle input to Torque. See Astrom and Murray 2008 Chapter 3.
-
-    :param omega: throttle
-    :type omega: float
-
-    :return: motor torque
-    :rtype: float
-    """
-    omega_m = parameters['omega_m']
-    t_m = parameters['T_m']
-    beta = parameters['beta']
-    return t_m*(1 - beta*(omega / omega_m - 1)**2)
-
-
-# pylint: disable-msg=too-many-locals
-def step_forward(state_vals, disturbances, desired_v, dt):
-    """
-    One-step Euler integrator that takes the current state, disturbances, and desired velocity and calculates the subsequent state.
-
-    :param state_vals: current state [position, velocity, road_angle]
-    :type state_vals: np.matrix
-
-    :param disturbances: current hill angle
-    :type disturbances: np.matrix
-
-    :param desired_velocity: current desired velocity
-    :type desired_velocity: np.matrix
-
-    :param dt: time step (seconds)
-    :type dt: float
-
-    :return: a tuple consisting of:
-            - new_state: new state
-            - u: control inputs
-    :rtype: tuple -> (np.matrix, np.matrix)
-    """
-    p = state_vals[0, -1]
[…]
