### 2022-01-05 766afe4
make {} default for options
