### 2022-01-07 4cc09c1
savgoldiff should have options=None
