### 2022-01-10 44cc0d3
make known_dynamics function more generic
