+++ pynumdiff/kalman_smooth/__init__.py
@@ -5,6 +5,5 @@ from pynumdiff.kalman_smooth._kalman_smooth import constant_velocity
 from pynumdiff.kalman_smooth._kalman_smooth import constant_acceleration
 from pynumdiff.kalman_smooth._kalman_smooth import constant_jerk
 from pynumdiff.kalman_smooth._kalman_smooth import known_dynamics
-from pynumdiff.kalman_smooth._kalman_smooth import __known_dynamics__
 from pynumdiff.kalman_smooth._kalman_smooth import __kalman_forward_filter__
 from pynumdiff.kalman_smooth._kalman_smooth import __kalman_backward_smooth__
