+++ docs/requirements.txt
@@ -46,7 +46,7 @@ pandocfilters==1.4.3
 parso==0.8.2
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.3.2
+Pillow==9.0.0
 pluggy==0.13.1
 prometheus-client==0.11.0
 prompt-toolkit==3.0.19
