### 2022-01-18 #3922 @christinahedges
Hello @florisvb and @pmli, thank you for your patience in getting this review started. Unfortunately, perhaps because of the holidays, I've struggled finding a reviewer for this submission despite quite a lot of digging. @florisvb if you have any further suggestions of unconflicted reviewers please do let me know otherwise I think we will have to be patient. I will also see if I can rope in some other JOSS editors so we can get this kicked off.

Thanks again for your patience!
