### 2022-01-18 #3922 @florisvb
Hi Christina,

What qualifications do you typically look for in JOSS reviewers? Should
they be people that hold academic positions, or are people in data science
/ software development positions also eligible? If the latter, I may be
able to come up with additional candidates.

Thanks!

- Floris

On Tue, Jan 18, 2022 at 9:34 AM Christina Hedges ***@***.***>
wrote:

> Hello @florisvb <https://github.com/florisvb> and @pmli
> <https://github.com/pmli>, thank you for your patience in getting this
> review started. Unfortunately, perhaps because of the holidays, I've
> struggled finding a reviewer for this submission despite quite a lot of
> digging. @florisvb <https://github.com/florisvb> if you have any further
> suggestions of unconflicted reviewers please do let me know otherwise I
> think we will have to be patient. I will also see if I can rope in some
> other JOSS editors so we can get this kicked off.
>
> Thanks again for your patience!
>
