### 2022-01-20 #3922 @christinahedges
If there are people in your field you feel have used the tool in their work, but are not conflicted with you, that would be ideal. Anyone in data science and software would still be totally fine, so long as it's clear they are an expert reviewer for some aspect of the tool. I will continue to ping people, thanks for your help!
