### 2022-01-20 #4078 @whedon
[REVIEW]: PyNumDiff: A Python package for numerical differentiation of noisy time-series data
**Submitting author:** @florisvb (<a href="http://orcid.org/0000-0001-6538-7179">Floris van Breugel</a>)
**Repository:** https://github.com/florisvb/PyNumDiff
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2.4
