### 2022-01-20 #4078 @christinahedges
@florisvb, @pmli , @billtubbs – This is the review thread for the PyNumDiff paper. Please don't hesitate to message me here if you have questions (use @christinahedges). ✉️

Please read the "Reviewer instructions & questions" in the first comment above to get started. If you get lost, you can also see the reviewer guidelines.

Both reviewers have checklists at the top of this thread (in that first comment) with the JOSS requirements. To review for JOSS, @pmli and @billtubbs will step through that checklist for PyNumDiff and assess the package. As you go over the submission, please check any items that you feel have been satisfied. If you are concerned about a requirement, please discuss it here on this thread 🧵 . Feel free to post about questions/concerns as they come up as you go through your review. Discussion between the authors/reviewers and myself is encouraged!

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention this issue (#4078) so that a link is created to this thread (and I can keep an eye on what is happening).

We aim for the review process to be completed within about 4-6 weeks but please make a start well ahead of this as JOSS reviews are by their nature iterative and any early feedback you may be able to provide to the author will be very helpful in meeting this schedule. When you're finished with your checklist, leave a comment and @ me to let everyone know your review is complete.

For this submission, don't forget there is a [published journal article](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7899139/) discussing the PyNumDiff algorithm. The JOSS editors have agreed that this JOSS submission is a separate and allowable submission given this published article. I'm leaving it here in case the reviewers find it useful during their review.
