+++ docs/requirements.txt
@@ -17,7 +17,7 @@ ecos==2.0.7.post1
 entrypoints==0.3
 iniconfig==1.1.1
 ipykernel==6.0.1
-ipython==7.25.0
+ipython==7.31.1
 ipython-genutils==0.2.0
 isort==5.9.1
 jedi==0.18.0
