### 2022-01-31 florisvb/PyNumDiff#11 @pmli
Installation issues
I found a few issues with the installation instructions in `README.md`:

1. It seems that the minimal required Python version is 3.5 (the `@` operator is used in `linear_model/_linear_model.py`).
2. Running `pip install pynumdiff` in a fresh virtual environment brings up issues with `pychebfun` for me. It seems that `numpy`, `scipy`, and `matplotlib` need to be installed prior to installing `pychebfun` (and `pynumdiff`).
3. Regarding installing from source, I heard somewhere that `pip install .` is recommended over `python setup.py install`. And, indeed, running the latter raises `SetuptoolsDeprecationWarning` and `EasyInstallDeprecationWarning` for me.

(This is related to my JOSS review: https://github.com/openjournals/joss-reviews/issues/4078)
