### 2022-02-01 cbbef67
remove required packages that are actually optional
