### 2022-02-01 bd0ade9
update readme with more install instructions
