### 2022-02-01 5a0caa6
additional install instructions
