### 2022-02-01 a683faf
clean up install instructions
