### 2022-02-01 florisvb/PyNumDiff#11 @florisvb
Thanks for identifying these issues. I have attempted to resolve all of them through various methods.

1. I removed the @ operator (replaced with cvxpy.matmul), though I also added python>=3.5 as a requirement, though in principle most of the code should work on python 2.7.
2. I removed pychebfun and cvxpy as requirements, and added them as optional requirements. This eliminates that particular issue, furthermore, pychebfun is only required for one option, and there is no need to require all users to have this installed.
3. It is challenging to meet all the installation requirements (especially on an older system like Ubuntu 16.04). To help with this I added more detailed install instructions include several examples that work through a complete install in an empty virtual environment.

Let me know if these changes resolve the issues raised.
