### 2022-02-01 florisvb/PyNumDiff#11 @pmli
Ok, thanks, I guess this resolves the issues, so I'll close this. I would just add some comments:

1. From using [`vermin`](https://github.com/netromdk/vermin), it seems that the minimum required Python version is 3.3. (I do not mean to suggest with this that PyNumDiff should support Python versions that had reached their end of life.)
2. Sounds good.
3. When I had been using Ubuntu 16.04, I had a good experience using [`pyenv`](https://github.com/pyenv/pyenv) to install newer versions of Python without admin privileges. Does this help you to avoid the installation instructions for these cases?
