### 2022-02-05 #3922 @billtubbs
Hi @christinahedges .  Just realized I asked a question to a bot (@whedon).  I have a problem in the #4708 review: See [this comment](https://github.com/openjournals/joss-reviews/issues/4078#issuecomment-1030690507).

Should this checklist be clickable by me?
