### 2022-02-07 e5647ef
add __version__ to imports and increase micro version to 2
