+++ pynumdiff/__init__.py
@@ -1,6 +1,7 @@
 """
 Import useful functions from all modules
 """
+from pynumdiff.__version__ import __version__
 from pynumdiff.finite_difference import first_order, second_order
 from pynumdiff.smooth_finite_difference import mediandiff, meandiff, gaussiandiff, \
     friedrichsdiff, butterdiff, splinediff
