### 2022-02-11 florisvb/PyNumDiff#13 @pmli
Issues when running tests
I noticed the following issues when running `pytest pynumdiff` (with Python 3.7.11 and PyNumDiff installed with `pip install -e .`):

- in `test_optimize.py`, the exception `TypeError: pi_control() got an unexpected keyword argument 'simdt'` is raised
- after fixing the above issue by removing `simdt=0.01`:
    - some tests fail if `cvxopt` is not installed (maybe `pip install -r requirements.txt` is required before running the tests?)
    - in `test_optimize.py`, both `test_chebydiff` and `test_template` fail
    - many `PendingDeprecationWarnings` are raised because of `np.matrix` usage (with NumPy 1.21.0)
    - a `DeprecationWarning` is raised by `scipy.sparse.linalg.cg` (with SciPy 1.7.0)

(This is related to my JOSS review: https://github.com/openjournals/joss-reviews/issues/4078)
