### 2022-02-11 florisvb/PyNumDiff#16 @pmli
Issues in examples
- `examples/1_basic_tutorial.ipynb`:
    - I got `x.shape == (400,)` instead of `x.shape == (401,)`
    - in section `1.2 Finite Difference: Second Order`, I suppose that `second_order` should be called instead of `first_order`
- `examples/2a_optimizing_parameters_with_dxdt_known.ipynb`:
    - same issue with `simdt` as in #13
    - removing `simdt` and replacing `dt = 0.1` by `dt = 0.01` makes things work, but the generated data looks very different compared to the pre-generated figures
    - similar issue with `1.2 Finite Difference: Second Order` as above
- `examples/2b_optimizing_parameters_with_dxdt_unknown.ipynb`:
    - similar issues as for the `2a` example

(This is related to my JOSS review: https://github.com/openjournals/joss-reviews/issues/4078)
