### 2022-02-16 7ae13f2
remove simdt from pi_control call - simdt was removed from pi_control in recent changes
