+++ pynumdiff/tests/test_optimize.py
@@ -20,13 +20,12 @@ from pynumdiff.utils import simulate
 # simulation
 noise_type = 'normal'
 noise_parameters = [0, 0.01]
-dt = 0.1
-simdt = 0.01
+dt = 0.01
 timeseries_length = 2
 problem = 'pi_control'
 x, x_truth, dxdt_truth, extras = simulate.__dict__[problem](timeseries_length,
                                                             noise_parameters=noise_parameters,
-                                                            dt=dt, simdt=0.01)
+                                                            dt=dt)
 cutoff_frequency = 0.1
 log_gamma = -1.6 * np.log(cutoff_frequency) - 0.71 * np.log(dt) - 5.1
 tvgamma = np.exp(log_gamma)
