### 2022-02-16 7a07c85
more elegant catching of cvxpy not installed errors
