### 2022-02-17 ca6e2ce
removed deprecated np.matrix
