+++ pynumdiff/kalman_smooth/_kalman_smooth.py
@@ -24,18 +24,18 @@ def __kalman_forward_update__(xhat_fm, P_fm, y, u, A, B, C, R, Q):
     :param Q:
     :return:
     """
-    I = np.matrix(np.eye(A.shape[0]))
-    gammaW = np.matrix(np.eye(A.shape[0]))
+    I = np.array(np.eye(A.shape[0]))
+    gammaW = np.array(np.eye(A.shape[0]))
 
-    K_f = P_fm*C.T*(C*P_fm*C.T + R).I
+    K_f = P_fm@C.T@np.linalg.pinv(C@P_fm@C.T + R)
 
-    xhat_fp = xhat_fm + K_f*(y - C*xhat_fm)
+    xhat_fp = xhat_fm + K_f@(y - C@xhat_fm)
 
-    P_fp = (I - K_f*C)*P_fm
+    P_fp = (I - K_f@C)@P_fm
 
-    xhat_fm = A*xhat_fp + B*u
+    xhat_fm = A@xhat_fp + B@u
 
-    P_fm = A*P_fp*A.T + gammaW*Q*gammaW.T
+    P_fm = A@P_fp@A.T + gammaW@Q@gammaW.T
 
     return xhat_fp, xhat_fm, P_fp, P_fm
 
@@ -53,24 +53,15 @@ def __kalman_forward_filter__(xhat_fm, P_fm, y, u, A, B, C, R, Q):
     :param Q:
     :return:
     """
-    assert isinstance(xhat_fm, np.matrix)
-    assert isinstance(P_fm, np.matrix)
-    assert isinstance(A, np.matrix)
-    assert isinstance(B, np.matrix)
-    assert isinstance(C, np.matrix)
-    assert isinstance(R, np.matrix)
-    assert isinstance(Q, np.matrix)
-    assert isinstance(y, np.matrix)
     if u is None:
-        u = np.matrix(np.zeros([B.shape[1], y.shape[1]]))
-    assert isinstance(u, np.matrix)
+        u = np.array(np.zeros([B.shape[1], y.shape[1]]))
 
     xhat_fp = None
     P_fp = []
     P_fm = [P_fm]
 
     for i in range(y.shape[1]):
-        _xhat_fp, _xhat_fm, _P_fp, _P_fm = __kalman_forward_update__(xhat_fm[:, -1], P_fm[-1], y[:, i], u[:, i],
+        _xhat_fp, _xhat_fm, _P_fp, _P_fm = __kalman_forward_update__(xhat_fm[:, [-1]], P_fm[-1], y[:, [i]], u[:, [i]],
                                                                      A, B, C, R, Q)
         if xhat_fp is None:
             xhat_fp = _xhat_fp
@@ -98,9 +89,9 @@ def __kalman_backward_smooth__(xhat_fp, xhat_fm, P_fp, P_fm, A):
     xhat_smooth = copy.copy(xhat_fp)
     P_smooth = copy.copy(P_fp)
     for t in range(N-2, -1, -1):
-        L = P_fp[t]*A.T*P_fm[t].I
-        xhat_smooth[:, t] = xhat_fp[:, t] + L*(xhat_smooth[:, t+1] - xhat_fm[:, t+1])
-        P_smooth[t] = P_fp[t] - L*(P_smooth[t+1] - P_fm[t+1])
+        L = P_fp[t]@A.T@np.linalg.pinv(P_fm[t])
+        xhat_smooth[:, [t]] = xhat_fp[:, [t]] + L@(xhat_smooth[:, [t+1]] - xhat_fm[:, [t+1]])
+        P_smooth[t] = P_fp[t] - L@(P_smooth[t+1] - P_fm[t+1])
 
[…]
