### 2022-02-17 77eb59a
elegantly fail and ignore tests when cvxpy not installed
