### 2022-02-17 178af1d
elegantly fail and ignore tests when cvxpy or pychebfun not isntalled
