### 2022-02-17 a7f76c6
skip tests when they require cvxpy or pychebfun and those are not installed
