+++ pynumdiff/tests/test_optimize.py
@@ -8,13 +8,22 @@ from unittest import TestCase
 from pynumdiff.optimize.finite_difference import first_order
 from pynumdiff.optimize.smooth_finite_difference import mediandiff, meandiff, gaussiandiff, \
     friedrichsdiff, butterdiff, splinediff
-from pynumdiff.optimize.total_variation_regularization import velocity, iterative_velocity, \
-    acceleration, smooth_acceleration, jerk, jerk_sliding
-from pynumdiff.optimize.linear_model import savgoldiff, spectraldiff, \
-    polydiff, chebydiff, lineardiff
+from pynumdiff.optimize.total_variation_regularization import *
+from pynumdiff.optimize.linear_model import *
 from pynumdiff.optimize.kalman_smooth import constant_velocity, constant_acceleration, \
     constant_jerk
 from pynumdiff.utils import simulate
+
+import logging as _logging
+
+_logging.basicConfig(
+    level=_logging.INFO,
+    format="%(asctime)s [%(levelname)s] %(message)s",
+    handlers=[
+        _logging.FileHandler("debug.log"),
+        _logging.StreamHandler()
+    ]
+)
 
 
 # simulation
@@ -30,100 +39,118 @@ cutoff_frequency = 0.1
 log_gamma = -1.6 * np.log(cutoff_frequency) - 0.71 * np.log(dt) - 5.1
 tvgamma = np.exp(log_gamma)
 
-
 class TestOPT(TestCase):
-    # def test_first_order(self):
-    #     params_1, val_1 = first_order(x, dt, params=None, options={'iterate': True},
-    #                                   tvgamma=tvgamma, dxdt_truth=dxdt_truth)
-    #     params_2, val_2 = first_order(x, dt, params=None, options={'iterate': True},
-    #                                   tvgamma=0, dxdt_truth=None)
-    #     self.assertListEqual(params_1, [12])
-    #     self.assertListEqual(params_2, [1])
-    #
-    # def test_mediandiff(self):
-    #     params_1, val_1 = mediandiff(x, dt, params=None, options={'iterate': True},
-    #                                  tvgamma=tvgamma, dxdt_truth=dxdt_truth)
-    #     params_2, val_2 = mediandiff(x, dt, params=None, options={'iterate': True},
-    #                                  tvgamma=0, dxdt_truth=None)
-    #     self.assertListEqual(params_1, [5, 5])
-    #     self.assertListEqual(params_2, [5, 5])
-    #
-    # def test_meandiff(self):
-    #     params_1, val_1 = meandiff(x, dt, params=None, options={'iterate': True},
-    #                                tvgamma=tvgamma, dxdt_truth=dxdt_truth)
-    #     params_2, val_2 = meandiff(x, dt, params=None, options={'iterate': True},
-    #                                tvgamma=0, dxdt_truth=None)
-    #     self.assertListEqual(params_1, [5, 1])
-    #     self.assertListEqual(params_2, [5, 1])
-    #
-    # def test_gaussiandiff(self):
-    #     params_1, val_1 = gaussiandiff(x, dt, params=None, options={'iterate': True},
-    #                                    tvgamma=tvgamma, dxdt_truth=dxdt_truth)
-    #     params_2, val_2 = gaussiandiff(x, dt, params=None, options={'iterate': True},
-    #                                    tvgamma=0, dxdt_truth=None)
-    #     self.assertListEqual(params_1, [5, 5])
-    #     self.assertListEqual(params_2, [5, 1])
-    #
-    # def test_friedrichsdiff(self):
-    #     params_1, val_1 = friedrichsdiff(x, dt, params=None, options={'iterate': True},
[…]
