### 2022-02-17 a162e21
fixed tests, skip tests that require cvxpy or pychebfun if those are not installed
