+++ pynumdiff/tests/test_optimize.py
@@ -45,7 +45,7 @@ class TestOPT(TestCase):
                                       tvgamma=tvgamma, dxdt_truth=dxdt_truth)
         params_2, val_2 = first_order(x, dt, params=None, options={'iterate': True},
                                       tvgamma=0, dxdt_truth=None)
-        self.assertListEqual(params_1, [12])
+        self.assertListEqual(params_1, [5])
         self.assertListEqual(params_2, [1])
     
     def test_mediandiff(self):
@@ -53,8 +53,8 @@ class TestOPT(TestCase):
                                      tvgamma=tvgamma, dxdt_truth=dxdt_truth)
         params_2, val_2 = mediandiff(x, dt, params=None, options={'iterate': True},
                                      tvgamma=0, dxdt_truth=None)
-        self.assertListEqual(params_1, [5, 5])
-        self.assertListEqual(params_2, [5, 5])
+        self.assertListEqual(params_1, [2, 7])
+        self.assertListEqual(params_2, [1, 1])
     
     def test_meandiff(self):
         params_1, val_1 = meandiff(x, dt, params=None, options={'iterate': True},
@@ -62,7 +62,7 @@ class TestOPT(TestCase):
         params_2, val_2 = meandiff(x, dt, params=None, options={'iterate': True},
                                    tvgamma=0, dxdt_truth=None)
         self.assertListEqual(params_1, [5, 1])
-        self.assertListEqual(params_2, [5, 1])
+        self.assertListEqual(params_2, [1, 1])
     
     def test_gaussiandiff(self):
         params_1, val_1 = gaussiandiff(x, dt, params=None, options={'iterate': True},
@@ -70,32 +70,35 @@ class TestOPT(TestCase):
         params_2, val_2 = gaussiandiff(x, dt, params=None, options={'iterate': True},
                                        tvgamma=0, dxdt_truth=None)
         self.assertListEqual(params_1, [5, 5])
-        self.assertListEqual(params_2, [5, 1])
+        self.assertListEqual(params_2, [1, 1])
     
     def test_friedrichsdiff(self):
         params_1, val_1 = friedrichsdiff(x, dt, params=None, options={'iterate': True},
                                          tvgamma=tvgamma, dxdt_truth=dxdt_truth)
         params_2, val_2 = friedrichsdiff(x, dt, params=None, options={'iterate': True},
                                          tvgamma=0, dxdt_truth=None)
-        self.assertListEqual(params_1, [5, 4])
-        self.assertListEqual(params_2, [5, 1])
+        self.assertListEqual(params_1, [9, 1])
+        self.assertListEqual(params_2, [1, 1])
     
     def test_butterdiff(self):
-        return
+        params_1, val_1 = butterdiff(x, dt, params=None, tvgamma=tvgamma, dxdt_truth=dxdt_truth)
+        params_2, val_2 = butterdiff(x, dt, params=None, tvgamma=0, dxdt_truth=None)
+        np.testing.assert_almost_equal(params_1, [9, 0.157], decimal=3)
+        np.testing.assert_almost_equal(params_2, [2, 0.99], decimal=3)
     
     def test_splinediff(self):
         params_1, val_1 = splinediff(x, dt, params=None, options={'iterate': True},
                                      tvgamma=tvgamma, dxdt_truth=dxdt_truth)
         params_2, val_2 = splinediff(x, dt, params=None, options={'iterate': True},
                                      tvgamma=0, dxdt_truth=None)
[…]
