### 2022-02-17 5a1f8fe
add atol to linalg.cg calls
