### 2022-02-17 8ccb970
removed np.matrix
