### 2022-02-18 28bc8aa
remove np.matrix
