+++ pynumdiff/utils/simulate.py
@@ -84,10 +84,10 @@ def sine(timeseries_length=4, noise_type='normal', noise_parameters=(0, 0.5), ra
     for f in frequencies:
         x += magnitude/len(frequencies)*_np.sin(t*2*_np.pi*f)
         dxdt += magnitude/len(frequencies)*_np.cos(t*2*_np.pi*f)*2*_np.pi*f
-    actual_vals = _np.matrix(_np.vstack((x, dxdt)))
+    actual_vals = _np.array(_np.vstack((x, dxdt)))
     noisy_x = __add_noise__(x, noise_type, noise_parameters, random_seed)
     #
-    noisy_measurements = _np.matrix(noisy_x)
+    noisy_measurements = _np.array(noisy_x)
     #
     noisy_pos = _np.ravel(noisy_measurements)
     pos = _np.ravel(actual_vals[0, :])
@@ -161,8 +161,8 @@ def triangle(timeseries_length=4, noise_type='normal', noise_parameters=(0, 0.5)
 
     noisy_x = __add_noise__(x, noise_type, noise_parameters, random_seed)
 
-    actual_vals = _np.matrix(_np.vstack((x, dxdt)))
-    noisy_measurements = _np.matrix(noisy_x)
+    actual_vals = _np.array(_np.vstack((x, dxdt)))
+    noisy_measurements = _np.array(noisy_x)
 
     noisy_pos = _np.ravel(noisy_measurements)
     pos = _np.ravel(actual_vals[0, :])
@@ -217,13 +217,13 @@ def pop_dyn(timeseries_length=4, noise_type='normal', noise_parameters=(0, 0.5),
         x.append(x[-1] + simdt*dxdt[-1])
         dxdt.append(r*x[-1]*(1-x[-1]/K))
 
-    x = _np.matrix(x)
-    dxdt = _np.matrix(dxdt)
+    x = _np.array(x)
+    dxdt = _np.array(dxdt)
 
     noisy_x = __add_noise__(x, noise_type, noise_parameters, random_seed)
 
-    actual_vals = _np.matrix(_np.vstack((x, dxdt)))
-    noisy_measurements = _np.matrix(noisy_x)
+    actual_vals = _np.array(_np.vstack((x, dxdt)))
+    noisy_measurements = _np.array(noisy_x)
 
     noisy_pos = _np.ravel(noisy_measurements)
     pos = _np.ravel(actual_vals[0, :])
@@ -270,12 +270,11 @@ def linear_autonomous(timeseries_length=4, noise_type='normal', noise_parameters
     """
     t = _np.arange(0, timeseries_length, simdt)
 
-    A = _np.matrix([[1, simdt, 0], [0, 1, simdt], [-100, -3, 0.01]])
-    print(A)
-    x0 = _np.matrix([[0], [2], [0]])
+    A = _np.array([[1, simdt, 0], [0, 1, simdt], [-100, -3, 0.01]])
+    x0 = _np.array([[0], [2], [0]])
     xs = x0
     for _ in t:
-        x = A*xs[:,-1]
+        x = A@xs[:,[-1]]
         xs = _np.hstack((xs, x))
 
     x = xs[0,:]
@@ -333,8 +332,8 @@ def pi_control(timeseries_length=4, noise_type='normal', noise_parameters=(0, 0.
     
     noisy_x = __add_noise__(x, noise_type, noise_parameters, random_seed)
 
-    actual_vals = _np.matrix(_np.vstack((x, dxdt)))
-    noisy_measurements = _np.matrix(noisy_x)
+    actual_vals = _np.array(_np.vstack((x, dxdt)))
+    noisy_measurements = _np.array(noisy_x)
 
[…]
