### 2022-02-18 fa94f22
warnings for optimizing with tvr and linear models
