### 2022-02-18 17e6d2d
make more robust to dealing with lack of cvxpy and pychebfun
