+++ pynumdiff/__init__.py
@@ -5,9 +5,8 @@ from pynumdiff.__version__ import __version__
 from pynumdiff.finite_difference import first_order, second_order
 from pynumdiff.smooth_finite_difference import mediandiff, meandiff, gaussiandiff, \
     friedrichsdiff, butterdiff, splinediff
-from pynumdiff.total_variation_regularization import velocity, iterative_velocity, \
-    acceleration, smooth_acceleration, jerk, jerk_sliding
-from pynumdiff.linear_model import savgoldiff, spectraldiff, polydiff, chebydiff, \
-    lineardiff
+from pynumdiff.total_variation_regularization import *
+from pynumdiff.linear_model import *
 from pynumdiff.kalman_smooth import constant_velocity, constant_acceleration, constant_jerk, \
     known_dynamics
+
