### 2022-02-18 florisvb/PyNumDiff#13 @florisvb
I fixed all of these in the most recent push.

Due to a previous issue regarding installing the package I had decided to make the code robust to installing and working as best as possible without installing cvxpy or pychebfun. I think this is the right move, because many of the functions (most, really) work perfectly without these packages, and I don't want to force users to jump through the hoops of installing either of these less common packages.

This does however cause issues when running tests without having these packages installed. My solution, perhaps unorthodox and I am open to suggestions, was to check if cvxpy and pychebfun are installed in the tests that require them, and skip those tests if those packages are not installed. I have logging.info messages for those cases, but pytest does not by default allow those to get printed to the terminal, so they skip silently. I would prefer a warning to show up, I'm open to suggestions there.

All tests pass in python 3.5.2 with cvxpy and pychebfun installed.

In python 3.7.9 without cvxpy or pychebfun installed all tests pass, though the tests that require cvxpy and pychebfun are skipped silently.
