### 2022-02-18 674210d
add .txt to LICENSE
