### 2022-02-18 florisvb/PyNumDiff#16 @florisvb
all issues fixed in latest push.
