### 2022-02-18 florisvb/PyNumDiff#13 @florisvb
Let me know what you think of my solution for the tests, if its okay I'll close, if not, if you have a suggestion for a better way to handle it that would be very helpful.
