### 2022-02-19 612cd21
make kalman smoother robust to missing data
