### 2022-02-23 florisvb/PyNumDiff#13 @pmli
> Due to a previous issue regarding installing the package I had decided to make the code robust to installing and working as best as possible without installing cvxpy or pychebfun. I think this is the right move, because many of the functions (most, really) work perfectly without these packages, and I don't want to force users to jump through the hoops of installing either of these less common packages.

That sounds good to me.

> This does however cause issues when running tests without having these packages installed. My solution, perhaps unorthodox and I am open to suggestions, was to check if cvxpy and pychebfun are installed in the tests that require them, and skip those tests if those packages are not installed. I have logging.info messages for those cases, but pytest does not by default allow those to get printed to the terminal, so they skip silently. I would prefer a warning to show up, I'm open to suggestions there.

I would suggest using [`pytest.skip`](https://docs.pytest.org/en/stable/how-to/skipping.html) instead of logging. That way, `pytest` will report how many tests were skipped.

> All tests pass in python 3.5.2 with cvxpy and pychebfun installed.

For me, with Python 3.7.11, one test fails:

```python
    def test_iterative_velocity(self):
        params_1, val_1 = iterative_velocity(x, dt, params=None, tvgamma=tvgamma, dxdt_truth=dxdt_truth)
        params_2, val_2 = iterative_velocity(x, dt, params=None, tvgamma=0, dxdt_truth=None)
>       self.assertListEqual(params_1, [2, 0.0001])
E       AssertionError: Lists differ: [1, 0.0001] != [2, 0.0001]
E
E       First differing element 0:
E       1
E       2
E
E       - [1, 0.0001]
E       ?  ^
E
E       + [2, 0.0001]
E       ?  ^

pynumdiff/tests/test_optimize.py:100: AssertionError
```
