### 2022-02-23 2dc66f4
use pytest.skip when cvxpy or pychebufun cannot be imported
