### 2022-02-23 f1993f9
attempt to make iterative velocity test more robust
