### 2022-02-23 florisvb/PyNumDiff#13 @pmli
> I replaced the logging calls with pytest.skip,

Great, it looks good.

> and I tried to make the test that failed for you more robust. Hard for me to fully debug though, since I get a different value (i.e. 2 instead of 1). What version of scipy and numpy do you have?

I used the versions from `requirements.txt`.
