### 2022-02-23 florisvb/PyNumDiff#13 @florisvb
That is odd, I'm using the same versions of scipy and numpy, on python 3.7.

In any case, could you rerun the tests with the latest code to see if it passes now?
