### 2022-02-24 florisvb/PyNumDiff#13 @pmli
Yes, the relaxed test passes.
