### 2022-02-24 florisvb/PyNumDiff#13 @florisvb
Excellent, thanks.

I believe everything from this issue is now addressed, so I'm going to close the issue. If I missed something let me know.
