### 2022-02-25 florisvb/PyNumDiff#16 @pmli
There is an issue in the first example (it runs fine on my machine):

![example1_issue](https://user-images.githubusercontent.com/1268421/155746540-c9eee462-a8e6-4908-b31e-2ec6ce192a52.png)
