### 2022-02-25 florisvb/PyNumDiff#17 @pmli
Docs issues
- The [readthedocs index page](https://pynumdiff.readthedocs.io/en/latest/index.html) doesn't seem to be synchronized with the information in the README.
- It looks like the first docstring line is the same for:
    - `pynumdiff.kalman_smooth._kalman_smooth.constant_acceleration`,
    - `pynumdiff.kalman_smooth._kalman_smooth.known_dynamics`, and
    - `pynumdiff.kalman_smooth._kalman_smooth.savgol_const_accel`.
- The documentation of [`_pi_cruise_control`](https://pynumdiff.readthedocs.io/en/latest/utils/_pi_cruise_control.html) seems to be showing `old_pi_cruise_control.py`.

(This is related to my JOSS review: https://github.com/openjournals/joss-reviews/issues/4078)
