### 2022-02-25 florisvb/PyNumDiff#18 @pmli
JOSS paper references
The paper content looks very good to me. I just have some minor comments regarding references:

- there is a double "Van" in `van2020numerical`,
- some uppercase letters are not protected in the titles of a few references ("kalman" and "savitzky-golay"),
- there is a more recent [SciPy paper](https://doi.org/10.1038/s41592-019-0686-2) that you could consider citing.

(This is related to my JOSS review: https://github.com/openjournals/joss-reviews/issues/4078)
