### 2022-02-28 a44179e
fix references and add new scipy ref
