+++ paper.md
@@ -35,7 +35,7 @@ Regardless of application domain, scientists of various levels of mathematical e
 
 # State of the field
 
-Currently, practitioners in need of numerical differentiation tools must often implement a number of methods themselves, before selecting one that is appropriate for their application. High-quality data can leverage computationally efficient and algorithmically simple methods such as the finite-difference, as implemented by standard packages such as NumPy [@harris2020array], SciPy [@scipy], or specialized packages life findiff [@findiff]. Data that are sparse and noisy, however, require more sophisticated algorithms that pracitioners must build themselves based on routines implemented across modules found in disparate packages such as SciPy, PyKalman [@pykalman], PyDMD [@demo18pydmd], or stand alone scripts such as these [implementations of total variation regularization](https://sites.google.com/site/dnartrahckcir/home/tvdiff-code) [@rudin1992nonlinear; @chartrand2011numerical]. At present, there is no centralized repository that offers a diverse range of vetted numerical differentiation tools under a unified API in Python, or other software languages. 
+Currently, practitioners in need of numerical differentiation tools must often implement a number of methods themselves, before selecting one that is appropriate for their application. High-quality data can leverage computationally efficient and algorithmically simple methods such as the finite-difference, as implemented by standard packages such as NumPy [@harris2020array], SciPy [@scipy][@Virtanen2020], or specialized packages life findiff [@findiff]. Data that are sparse and noisy, however, require more sophisticated algorithms that pracitioners must build themselves based on routines implemented across modules found in disparate packages such as SciPy, PyKalman [@pykalman], PyDMD [@demo18pydmd], or stand alone scripts such as these [implementations of total variation regularization](https://sites.google.com/site/dnartrahckcir/home/tvdiff-code) [@rudin1992nonlinear; @chartrand2011numerical]. At present, there is no centralized repository that offers a diverse range of vetted numerical differentiation tools under a unified API in Python, or other software languages. 
 
 	
 # Summary
