+++ paper.bib
@@ -1,6 +1,6 @@
 @article{van2020numerical,
   title={Numerical differentiation of noisy data: A unifying multi-objective optimization framework},
-  author={Van Van Breugel, Floris and Kutz, J Nathan and Brunton, Bingni W},
+  author={Van Breugel, Floris and Kutz, J Nathan and Brunton, Bingni W},
   journal={IEEE Access},
   volume={8},
   pages={196865--196877},
@@ -72,7 +72,7 @@
 }
 
 @article{schafer2011savitzky,
-  title={What is a Savitzky-Golay filter?[lecture notes]},
+  title={What is a {Savitzky-Golay} filter?[lecture notes]},
   author={Schafer, Ronald W},
   journal={IEEE Signal processing magazine},
   volume={28},
@@ -103,7 +103,7 @@
 }
 
 @article{henderson2010fundamentals, 
-title={Fundamentals of Kalman Filtering: A Practical Approach}, 
+title={Fundamentals of {Kalman} Filtering: A Practical Approach}, 
 volume={105}, 
 number={1049}, 
 journal={The Aeronautical Journal}, 
@@ -116,7 +116,7 @@ doi={10.1017/S000192400001232X},
 }
 
 @article{aravkin2017generalized,
-  title={Generalized Kalman smoothing: Modeling and algorithms},
+  title={Generalized {Kalman} smoothing: Modeling and algorithms},
   author={Aravkin, Aleksandr and Burke, James V and Ljung, Lennart and Lozano, Aurelie and Pillonetto, Gianluigi},
   journal={Automatica},
   volume={86},
@@ -163,9 +163,23 @@ doi={10.1017/S000192400001232X},
   url = "http://www.scipy.org/"
 }
 
+@article{Virtanen2020,
+  doi = {10.1038/s41592-019-0686-2},
+  url = {https://doi.org/10.1038/s41592-019-0686-2},
+  year = {2020},
+  month = feb,
+  publisher = {Springer Science and Business Media {LLC}},
+  volume = {17},
+  number = {3},
+  pages = {261--272},
[…]
