### 2022-03-01 florisvb/PyNumDiff#17 @florisvb
I fixed issues with the documentation text itself.

I am still working on fixing the readthedocs issue, which is due to switching the github repo that the project is primarily being maintained on.
