+++ docs/source/conf.py
@@ -24,7 +24,7 @@ copyright = '2021, Floris van Breugel, Yuying Liu'
 author = 'Floris van Breugel, Yuying Liu'
 
 # The full version, including alpha/beta/rc tags
-release = '0.0.1'
+release = '0.1.2'
 
 
 # -- General configuration ---------------------------------------------------
@@ -55,3 +55,5 @@ html_theme = 'sphinx_rtd_theme'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+
+master_doc = 'index'
