+++ docs/source/index.rst
@@ -1,111 +1,298 @@
-Welcome to PyNumDiff's documentation!
-=====================================
+PyNumDiff
+=========
 
+Python methods for numerical differentiation of noisy data, including
+multi-objective optimization routines for automated parameter selection.
 
-[Py]thon [Num]erical [Diff]erentiation for noisy time series measurements.
+.. raw:: html
 
-Description
------------
-PyNumDiff is a Python package that implements various methods for computing numerical derivatives of noisy data, which
-can be a critical step in developing dynamic models or designing control. There are four different families of methods
-implemented in this repository: smoothing followed by finite difference calculation, local approximation with linear
-models, Kalman filtering based methods and total variation regularization methods. Most of these methods have multiple
-parameters involved to tune. We take a principled approach and propose a multi-objective optimization framework for
-choosing parameters that minimize a loss function to balance the faithfulness and smoothness of the derivative estimate.
-For more details, refer to `this paper`_.
+   <p align="center">
 
+.. raw:: html
 
-Installation
-------------
-Prerequisites
-^^^^^^^^^^^^^
-PyNumDiff requires common packages like `numpy`, `scipy`, `matplotlib`, `pytest` (for unittests), `pylint`
-(for PEP8 style check). For a full list, you can check the file `requirements.txt`_
+   </p>
 
-In addition, it also requires certain additional packages:
+.. raw:: html
 
-- Total Variation Regularization methods: `cvxpy`_
-- Linear Model Chebychev: `pychebfun`_
+   <p align="center">
 
-When using `cvxpy`, our default solver is set to be `MOSEK` (highly recommended), you would need to download their
-free academic license from their website_. Otherwise, you can also
-use other solvers which are listed here_.
-
-The code is compatible with Python 3.x. It can be installed using pip or directly from the source code.
-
-
-Installing via PIP
-^^^^^^^^^^^^^^^^^^
-Mac and Linux users can install pre-built binary packages using pip.
-To install the package just type:
 ::
 
-    pip install pynumdiff
+    <a href="#travis" alt="Travis Build Status">
+        <img src="https://travis-ci.com/luckystarufo/PyNumDiff.svg?branch=upgrade"/></a>
+    <a href='https://pynumdiff.readthedocs.io/en/latest/?badge=latest'>
+        <img src='https://readthedocs.org/projects/pynumdiff/badge/?version=latest' alt='Documentation Status' /></a>
+    <a href="https://badge.fury.io/py/pynumdiff">
+        <img src="https://badge.fury.io/py/pynumdiff.svg" alt="PyPI version" height="18"></a>
+    <a href="https://zenodo.org/badge/latestdoi/361922468">
+        <img src="https://zenodo.org/badge/361922468.svg" alt="DOI"></a>
+    <a href="https://app.fossa.com/projects/git%2Bgithub.com%2Fflorisvb%2FPyNumDiff?ref=badge_shield">
+        <img src="https://app.fossa.com/api/projects/git%2Bgithub.com%2Fflorisvb%2FPyNumDiff.svg?type=shield" alt="DOI"></a>
 
-Installing from source
-^^^^^^^^^^^^^^^^^^^^^^
-The official distribution is on GitHub, and you can clone the repository using
-::
+.. raw:: html
 
-    git clone https://github.com/luckystarufo/PyNumDiff.git
+   </p>
 
[…]
