### 2022-03-02 b9a68e9
cvxpy not explicitly needed for optimize only certain functions
