+++ README.md
@@ -98,7 +98,6 @@ PyNumDiff requires common packages like `numpy`, `scipy`, `matplotlib`, `pytest`
 In addition, it also requires certain additional packages for select functions, though these are not required for a successful install of PyNumDiff:
 * Total Variation Regularization methods: [`cvxpy`](http://www.cvxpy.org/install/index.html)
 * Linear Model Chebychev: [`pychebfun`](https://github.com/pychebfun/pychebfun/)
-* Optimize: [`cvxpy`](http://www.cvxpy.org/install/index.html)
 
 When using `cvxpy`, our default solver is set to be `MOSEK` (highly recommended), you would need to download their 
 free academic license from their [website](https://www.mosek.com/products/academic-licenses/). Otherwise, you can also 
