### 2022-03-02 florisvb/PyNumDiff#17 @florisvb
The readthedocs index page is now synchronized with the README.

Closing.
