+++ pynumdiff/__version__.py
@@ -42,7 +42,7 @@ MAINTAINER = "Floris van Breugel, Yuying Liu"
 MAINTAINER_EMAIL = "fvanbreugel@unr.edu, yliu814@uw.edu"
 DESCRIPTION = description
 LONG_DESCRIPTION = long_description
-URL = "https://github.com/luckystarufo/PyNumDiff/tree/upgrade"
+URL = "https://github.com/florisvb/PyNumDiff"
 DOWNLOAD_URL = ""
 LICENSE = "MIT"
 AUTHOR = "Floris van Breugel, Yuying Liu"
