### 2022-03-03 e07d8ee
remove unnecessary usage section, try to fix list formatting, fix notebook links
