### 2022-03-03 834d83a
make tests more robust for travis
