+++ docs/source/index.rst
@@ -254,7 +254,7 @@ License
 
 |FOSSA Status|
 
-This project utilizes the `MIT LICENSE <LICENSE.txt>`__. 100%
+This project utilizes the `MIT LICENSE <https://github.com/florisvb/PyNumDiff/blob/master/LICENSE.txt>`__. 100%
 open-source, feel free to utilize the code however you like.
 
 .. |FOSSA Status| image:: https://app.fossa.com/api/projects/git%2Bgithub.com%2Fflorisvb%2FPyNumDiff.svg?type=large
