### 2022-03-03 3bea40e
remove redundant license link
