### 2022-03-04 #4078 @pmli
I believe I'm done with the review. All of the [issues](https://github.com/florisvb/PyNumDiff/issues?q=author%3Apmli+) I opened are closed.
