### 2022-03-04 #4078 @florisvb
Thanks for your patience and attention to detail, Petar, I really
appreciate it!

- Floris

On Fri, Mar 4, 2022 at 8:30 AM Petar Mlinarić ***@***.***>
wrote:

> I believe I'm done with the review. All of the issues
> <https://github.com/florisvb/PyNumDiff/issues?q=author%3Apmli+> I opened
> are closed.
>
