### 2022-03-09 #4078 @florisvb
- Floris

On Fri, Mar 4, 2022 at 2:35 PM Floris van Breugel ***@***.***>
wrote:

> Thanks for your patience and attention to detail, Petar, I really
> appreciate it!
>
> - Floris
>
> On Fri, Mar 4, 2022 at 8:30 AM Petar Mlinarić ***@***.***>
> wrote:
>
>> I believe I'm done with the review. All of the issues
>> <https://github.com/florisvb/PyNumDiff/issues?q=author%3Apmli+> I opened
>> are closed.
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/4078#issuecomment-1059317856>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AAB4EPB4EYNKD2BY7HACZD3U6I3D7ANCNFSM5MNTPQFQ>
>> .
>> Triage notifications on the go with GitHub Mobile for iOS
>> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
>> or Android
>> <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>>
>> You are receiving this because you were mentioned.Message ID:
>> ***@***.***>
>>
>
>
> --
> Floris van Breugel   |   http://www.florisvanbreugel.com
> Assistant Professor of Mechanical Engineering
> & Graduate Program for Neuroscience
> University of Nevada, Reno
>
> Wildlife and Landscape Photography
> Galleries: http://www.ArtInNaturePhotography.com/
> Blog: http://www.ArtInNaturePhotography.com/wordpress/
>

--
Floris van Breugel   |   http://www.florisvanbreugel.com
Assistant Professor of Mechanical Engineering
& Graduate Program for Neuroscience
University of Nevada, Reno

Wildlife and Landscape Photography
Galleries: http://www.ArtInNaturePhotography.com/
Blog: http://www.ArtInNaturePhotography.com/wordpress/
