### 2022-03-09 #4078 @florisvb
On Tue, Mar 8, 2022 at 4:36 PM The Open Journals editorial robot <
***@***.***> wrote:

> I'm sorry human, I don't understand that. You can see what commands I
> support by typing:
>
> @editorialbot commands
>
