### 2022-03-09 #4078 @florisvb
@christinahedges - I just wanted to check in to see where we are at with
the review. I resolved all of Petar's issues, but have not gotten any tasks
from Bill Tubbs.

I am new to this type of review process, so I apologize if I have missed
something.

- Floris

On Tue, Mar 8, 2022 at 4:37 PM The Open Journals editorial robot <
***@***.***> wrote:

> Hello @florisvb <https://github.com/florisvb>, here are the things you
> can ask me to do:
>
>
> # List all available commands
> @editorialbot commands
>
> # Get a list of all editors's GitHub handles
> @editorialbot list editors
>
> # Check the references of the paper for missing DOIs
> @editorialbot check references
>
> # Perform checks on the repository
> @editorialbot check repository
>
> # Adds a checklist for the reviewer using this command
> @editorialbot generate my checklist
>
> # Set a value for branch
> @editorialbot set joss-paper as branch
>
> # Generates the pdf paper
> @editorialbot generate pdf
>
> # Get a link to the complete list of reviewers
> @editorialbot list reviewers
>
>
