### 2022-03-10 #4078 @billtubbs
Hi @christinahedges, yes sorry, I was just waiting to do some more testing of one of the methods on some data I have from my current research.  I have done this now.  I probably don't have time to test all the methods but I will run their example notebooks on my computer just to make sure there are no bugs.  I haven't found any problems, bugs, or other improvement suggestions at all yet so I am likely to check all the boxes above.
