### 2022-03-11 #4078 @christinahedges
Hi @billtubbs, thank you that sounds great. Please let us know here or on an issue on the PyNumDiff repo if you do find any issues to discuss, I'll check back in when you've had a little more time. @florisvb it seems like we are nearly there!
