### 2022-03-15 #4078 @christinahedges
@billtubbs and @pmli, thank you so much for your thorough review of this submission! JOSS only works with great reviewers so thank you for your time and effort!

@florisvb it looks like the reviewers have completed their checklists, and your submission paper looks good to me. Please can you make a tagged release and archive, and report the version number and archive DOI in this thread. I will then recommend this submission be accepted.
