### 2022-03-18 #4078 @florisvb
Hi @christinahedges and @editorialbot - here are the version and DOI:

tagged release version number: PyNumDiff 0.1.2.3
archive DOI: DOI: 10.5281/zenodo.6366705
<https://doi.org/10.5281/zenodo.6366705>

- Floris

On Tue, Mar 15, 2022 at 2:03 PM Christina Hedges ***@***.***>
wrote:

> @billtubbs <https://github.com/billtubbs> and @pmli
> <https://github.com/pmli>, thank you so much for your thorough review of
> this submission! JOSS only works with great reviewers so thank you for your
> time and effort!
>
> @florisvb <https://github.com/florisvb> it looks like the reviewers have
> completed their checklists, and your submission paper looks good to me.
> Please can you make a tagged release and archive, and report the version
> number and archive DOI in this thread. I will then recommend this
> submission be accepted.
>
