+++ pynumdiff/__version__.py
@@ -45,8 +45,8 @@ LONG_DESCRIPTION = long_description
 URL = "https://github.com/florisvb/PyNumDiff"
 DOWNLOAD_URL = ""
 LICENSE = "MIT"
-AUTHOR = "Floris van Breugel, Yuying Liu"
-AUTHOR_EMAIL = "florisvb@gmail.com, liuyuyingufo@gmail.com"
+AUTHOR = "Floris van Breugel, Yuying Liu, Bingni W. Brunton, J. Nathan Kutz"
+AUTHOR_EMAIL = "florisvb@gmail.com, liuyuyingufo@gmail.com, bbrunton@uw.edu, kutz@uw.edu"
 PLATFORMS = "OS Independent"
 MAJOR = _version_major
 MINOR = _version_minor
