### 2022-03-21 #4078 @florisvb
@christinahedges -- do you mean in the pynumdiff/__version__.py  file? I
updated/corrected that author list. If there is a different list you were
referring to let me know.

Updated release info:

Tagged release: PyNumDiff 0.1.2.4
DOI: 10.5281/zenodo.6374098
Link: https://zenodo.org/record/6374098#.YjjGEVNKj6o

On Mon, Mar 21, 2022 at 10:41 AM Christina Hedges ***@***.***>
wrote:

> @florisvb <https://github.com/florisvb> thank you for adding this, it
> looks like the authors are not the same as in the paper (I think two author
> names may be missing?) Can you edit the meta data to make the authors
> consistent? Ideally they should be names not usernames.
>
