### 2022-03-21 #4078 @christinahedges
Hi @florisvb I actually meant the zenodo meta data:

You can see the author names under the title and they are usernames, not people names, and I believe there were 5 authors in your submission, you should be able to edit this meta data in zenodo.
