### 2022-03-21 #4078 @danielskatz
In addition, I've suggested a bunch of small changes in https://github.com/florisvb/PyNumDiff/pull/27 - please merge this or let me know what you disagree with.
