### 2022-03-21 #4078 @florisvb
Pull request accepted, and title on zenodo is now updated.

- Floris

On Mon, Mar 21, 2022 at 2:41 PM Daniel S. Katz ***@***.***>
wrote:

> In addition, I've suggested a bunch of small changes in
> florisvb/PyNumDiff#27 <https://github.com/florisvb/PyNumDiff/pull/27> -
> please merge this or let me know what you disagree with.
>
