### 2022-03-21 #4078 @danielskatz
sorry @florisvb - I also see some bib changes that are needed - in https://github.com/florisvb/PyNumDiff/pull/28.  And once this is done, I think I will need to talk to @arfon about some bugs in our processing of some of these elements...
