### 2022-03-21 #4078 @florisvb
Pull request accepted, thanks for catching those..

On Mon, Mar 21, 2022 at 4:00 PM Daniel S. Katz ***@***.***>
wrote:

> sorry @florisvb <https://github.com/florisvb> - I also see some bib
> changes that are needed - in florisvb/PyNumDiff#28
> <https://github.com/florisvb/PyNumDiff/pull/28>. And once this is done, I
> think I will need to talk to @arfon <https://github.com/arfon> about some
> bugs in our processing of some of these elements...
>
