### 2022-03-22 0807a56
update citation with JOSS doi and bibtex
