### 2022-10-11 #4078 @florisvb
On Mon, Mar 21, 2022 at 3:47 PM The Open Journals editorial robot <
***@***.***> wrote:

> 👋 @openjournals/joss-eics
> <https://github.com/orgs/openjournals/teams/joss-eics>, this paper is
> ready to be accepted and published.
>
> Check final proof 👉 openjournals/joss-papers#3077
> <https://github.com/openjournals/joss-papers/pull/3077>
>
> If the paper PDF and the deposit XML files look good in
> openjournals/joss-papers#3077
> <https://github.com/openjournals/joss-papers/pull/3077>, then you can now
> move forward with accepting the submission by compiling again with the
> command @editorialbot accept
>
